Update search results with use api.
